Move test helper into own file.
